opti & imports

+++ TP3/common.c
@@ -6,21 +6,21 @@ int intervalle_ms(struct timeval debut, struct timeval fin) {
 }
 
 int nb_ms_vers_centiemes(int nb_ms) {
-    int tmp = (int)(nb_ms / 10) % 100;
+    short int tmp = (short int)(nb_ms / 10) % 100;
     return tmp < 0 ? tmp * -1 : tmp;
 }
 
 int nb_ms_vers_secondes(int nb_ms) {
-    int tmp = (int)(nb_ms / 1000) % 60;
+    short int tmp = (short int)(nb_ms / 1000) % 60;
     return tmp < 0 ? tmp * -1 : tmp;
 }
 
 int nb_ms_vers_minutes(int nb_ms) {
-    int tmp = (int)(nb_ms / 60000) % 60;
+    short int tmp = (short int)(nb_ms / 60000) % 60;
     return tmp < 0 ? tmp * -1 : tmp;
 }
 
 int nb_ms_vers_heures(int nb_ms) {
-    int tmp = (int)(nb_ms / 360000);
+    short int tmp = (short int)(nb_ms / 360000);
     return tmp < 0 ? tmp * -1 : tmp;
 }
